moving type_utils.py out of the de folder so it can more naturally be used elsewhere
