+++ qiskit/providers/aer/pulse_new/type_utils.py
